feat: implement user exercise details submission with validation and caching
